# server/openapi.yaml
openapi: 3.0.3
info:
  title: ProjectHelper Timeline API
  description: |
    프로젝트 타임라인(간트 차트) 데이터를 조작하는 REST API.
    AI 에이전트가 작업(Task) 단위로 일정을 읽고 수정할 수 있도록 설계됨.

    ## 동시성 (리비전)
    - 모든 변경 응답에 `revision`(정수)이 포함된다. 데이터가 바뀔 때마다 1씩 증가.
    - 변경 요청에 `If-Match: <revision>` 헤더를 보내면 서버 리비전과 다를 때 409를 반환한다.
    - 헤더를 생략하면 무조건 쓰기(last-write-wins) — 기존 클라이언트 하위호환.
    - 브라우저 앱은 10초마다 `GET /api/revision`을 폴링해 외부 변경을 자동 반영한다.

    ## 인증 (계정이 있을 때만)
    - 서버에 계정이 하나도 없으면 `open` 모드다 — 인증 없이 지금까지처럼 동작한다.
    - 첫 관리자를 만든 뒤에는 모든 경로에 신원이 필요하다. 예외는 정확히
      `/`, `/guide`, `/openapi.yaml`, `/health`, `/auth/*` 뿐이다.
    - 사람은 세션 쿠키(`POST /auth/login`), **에이전트는 서비스 토큰**을 쓴다:
      `Authorization: Bearer <token>` (서버의 `PH_API_TOKENS` 에 `이름:역할:토큰` 으로 등록).
      감사 로그의 actor 는 그 이름이 되므로 사람 계정을 빌려 쓰지 않는다.
    - 역할은 viewer(읽기) ⊂ editor(쓰기) ⊂ admin(계정·프로젝트 삭제).
      권한이 모자라면 403, 신원이 없으면 401.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 로컬 API 서버 (dev는 vite proxy 경유 :5173/api)
  - url: https://{host}/api
    description: Caddy HTTPS 프록시 (basicauth 필요)
    variables:
      host:
        default: localhost
security:
  - serviceToken: []
  - {}   # 계정이 없는 배포(open 모드)에서는 인증 없이 호출된다
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      description: PH_API_TOKENS 에 등록된 서비스 토큰. 사람 사용자는 세션 쿠키를 쓴다.
    session:
      type: apiKey
      in: cookie
      name: ph_session
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema: { type: string }
      description: 기대하는 현재 리비전. 불일치 시 409.
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error: { type: string }
        revision: { type: integer }
    TimeRange:
      type: object
      properties:
        id: { type: string }
        startDate: { type: string, format: date, example: "2026-07-21" }
        endDate: { type: string, format: date }
        dependencies:
          type: array
          items: { type: string }
          description: 선행 엔티티 ID 목록 (range/milestone/task ID)
        color: { type: string, nullable: true, example: "#4A90E2" }
        label: { type: string }
    Milestone:
      type: object
      properties:
        id: { type: string }
        date: { type: string, format: date }
        label: { type: string }
        color: { type: string, example: "#5CB85C" }
        shape: { type: string, enum: [diamond, circle, triangle, square, star, flag] }
        labelPosition:
          type: string
          enum: [auto, top, bottom, left, right]
          description: auto(기본)는 서로 겹치지 않게 자동 배치. 손으로 고른 값은 그 자리를 먼저 차지한다.
        dependencies:
          type: array
          items: { type: string }
          description: 선행 엔티티 ID 목록 (range/milestone ID)
    Task:
      type: object
      description: 재귀 트리 노드. 날짜 정보는 timeRanges가 원본이며 startDate/endDate는 파생 캐시.
      properties:
        id: { type: string, example: "task-1784537581275-fe7oo66" }
        name: { type: string }
        timeRanges:
          type: array
          items: { $ref: '#/components/schemas/TimeRange' }
        color: { type: string }
        description: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100, description: 진행률 % }
        children:
          type: array
          items: { $ref: '#/components/schemas/Task' }
        expanded: { type: boolean }
        labels:
          type: array
          items: { type: string }
        parentId: { type: string, nullable: true }
        milestones:
          type: array
          items: { $ref: '#/components/schemas/Milestone' }
        dependencies:
          type: array
          items: { type: string }
          description: 레거시 — 로드 시 첫 timeRange로 마이그레이션되어 항상 빈 배열
        divider:
          type: object
          properties:
            enabled: { type: boolean }
            thickness: { type: integer }
            style: { type: string }
            color: { type: string }
  responses:
    NotFound:
      description: 대상 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 리비전 불일치 (If-Match)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /projects:
    get:
      summary: 프로젝트 목록 (각 프로젝트는 독립된 작업 트리/리비전/스냅샷)
      responses:
        '200':
          description: '{ ok, projects: [{id, name, owner, createdAt, updatedAt}] }'
    post:
      summary: 프로젝트 생성 — 새 일정 계획은 새 프로젝트에 작성 권장
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201': { description: '{ ok, project }' }
  /projects/{pid}:
    parameters:
      - name: pid
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: 프로젝트 이름 변경
      responses:
        '200': { description: '{ ok, project }' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: 프로젝트 삭제 (마지막 프로젝트는 400, 데이터는 _trash로 보존)
      responses:
        '200': { description: '{ ok }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  # 아래 모든 경로는 두 형태로 제공된다:
  #   /projects/{pid}/...  — 프로젝트 스코프 (권장)
  #   /...                 — default 프로젝트 별칭 (하위호환)
...
  /data:
    get:
      summary: 전체 작업 트리 조회 (통짜 블롭)
      responses:
        '200':
          description: OK. 데이터 없으면 data=null
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  revision: { type: integer }
                  data:
                    type: array
                    nullable: true
                    items: { $ref: '#/components/schemas/Task' }
    post:
      summary: 전체 작업 트리 교체 (통짜 블롭 — 주의해서 사용)
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/Task' }
      responses:
        '200': { description: 'OK { ok, revision }' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /snapshots:
    get:
      summary: 스냅샷(이름 지정 백업) 목록
      responses:
        '200': { description: '{ ok, data: [{id, name, date, data}] }' }
    post:
      summary: 스냅샷 저장 (대량 편집 전 백업 권장)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, data]
              properties:
                name: { type: string }
                data: { type: array, items: { $ref: '#/components/schemas/Task' } }
      responses:
        '200': { description: '{ ok, snapshot }' }
...
  /settings:
    get:
      summary: 전역 뷰 설정 조회
      description: >
        **프로젝트 스코프 밖이다** — 서버에 하나뿐인 전역 blob 이고
        `/projects/{pid}/settings` 는 없다. 다크모드·시간축·확대율 같은
        화면 표시 설정만 들어 있으며 일정 데이터는 없다.
      responses:
        '200': { description: '{ ok, data: {키: 스칼라} }' }
    post:
      summary: 전역 뷰 설정 변경 (보낸 키만 **병합**)
      description: >
        보내지 않은 키는 그대로 남는다 — 통짜 덮어쓰기였을 때, 설정 하나만 담은
        요청이 사용자의 나머지 뷰 설정을 전부 지웠다. 값은 문자열·숫자·불리언·null
        만 허용하고(중첩 객체·배열은 400), 키는 `[A-Za-z0-9_]{1,40}`, 최대 64개다.
        키 목록은 검증하지 않는다: 클라이언트 settingsStore.js 가 유일한 목록이고
        그것을 서버에 베껴 두면 설정을 추가할 때마다 두 곳이 갈라진다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                nullable: true
                oneOf:
                  - { type: string, maxLength: 200 }
                  - { type: number }
                  - { type: boolean }
            example: { darkMode: true, timeScale: week }
      responses:
        '200': { description: '{ ok, data: 병합된 전체 설정 }' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /events:
    get:
      summary: 감사 로그 — 이 프로젝트의 쓰기 이력(최신순, append-only)
      description: >
        모든 트리/스냅샷 쓰기가 한 줄씩 남는다. 되돌리기 용도가 아니라 추적용이다
        ("누가 언제 몇 개를 지웠는가"). 복구는 스냅샷과 세대 백업이 담당한다.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        '200':
          description: '{ ok, data: [{ts, actor, op, revision, nodes, prevNodes}] }'
...